For (some) consistency, make monster::foe an unsigned short, since that matches player::pet_target, and because tags.cc seems to save monster::foe as a short anyway.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@5461 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/externs.h
@@ -1028,7 +1028,7 @@ public:
     monster_spells spells;
     mon_attitude_type attitude;
     beh_type behaviour;
-    unsigned int foe;
+    unsigned short foe;
     char ench_countdown;
     mon_enchant_list enchantments;
     unsigned long flags;               // bitfield of boolean flags
